% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{assocfun}
\alias{assocfun}
\title{Association Function}
\usage{
assocfun(xind, yind, data)
}
\arguments{
\item{xind}{column index of the x-variable}

\item{yind}{column index of the y-variable}

\item{data}{data frame from which to pull the variables.}
}
\value{
a squared correlation.
}
\description{
Calculates the R-squared from a LOESS regression of
y on x.  Can be used with \code{outer} to produce the
a non-parametric correlation matrix.
}
