% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{boot_imp}
\alias{boot_imp}
\title{Bootstrap Importance Function}
\usage{
boot_imp(data, inds, obj)
}
\arguments{
\item{data}{A data frame}

\item{inds}{Indices to be passed into the function.}

\item{obj}{An object of class \code{lm}.}
}
\value{
A vector of standard deviation of predictions for
each term in the model.
}
\description{
Function to calculate bootstrap measures of importance.
This function must be passed to the \code{boot} function.
}
