% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{caption}
\alias{caption}
\title{Caption Grob}
\usage{
caption(lab, x = 0.5, y = 1, hj = 0.5, vj = 1, cx = 1, fs = 12, ft = "Arial")
}
\arguments{
\item{lab}{Text giving the caption text.}

\item{x}{Scalar giving the horizontal position of the label in \code{[0,1]}.}

\item{y}{Scalar giving the vertical position of the label in \code{[0,1]}.}

\item{hj}{Scalar giving horizontal justification parameter.}

\item{vj}{Scalar giving vertical justification parameter.}

\item{cx}{Character expansion factor}

\item{fs}{Font size}

\item{ft}{Font type}
}
\value{
A text grob.
}
\description{
Create a caption grob
}
