% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{cld.ss}
\alias{cld.ss}
\title{Compact Letter Display for Simple Slopes}
\usage{
\method{cld}{ss}(object, ..., level = 0.05)
}
\arguments{
\item{object}{An object of class `ss`}

\item{...}{Other arguments to be passed to generic function.}

\item{level}{Confidence level used for the letters.}
}
\value{
A compact letter matrix
}
\description{
Calculates a letter matrix for a simple-slopes output.
}
