% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{letter_plot}
\alias{letter_plot}
\title{Dot Plot with Leter Display}
\usage{
letter_plot(fits, letters)
}
\arguments{
\item{fits}{Output from \code{ggpredict} from the \pkg{ggeffects}}

\item{letters}{A matrix of character strings giving the letters from a
compact letter display.  This is most often from a call to \code{cld} from the
\pkg{multcomp} package.}
}
\description{
Produces an dot plot with error bars along with a compact letter display
}
