% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{lsa}
\alias{lsa}
\title{Linear Scatterplot Array}
\usage{
lsa(formula, xlabels = NULL, ylab = NULL, data)
}
\arguments{
\item{formula}{Formula giving the variables to be plotted.}

\item{xlabels}{Vector of character strings giving the labs of
variables to be used in place of the variable names.}

\item{ylab}{Character string giving y-variable label to be
used instead of variable name.}

\item{data}{A data frame that holds the variables to be plotted.}
}
\value{
A \code{cowplot} object.
}
\description{
Produces a linear scatterplot array with marginal histograms
}
