% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{normBand}
\alias{normBand}
\title{Kernel Density with Normal Density Overlay}
\usage{
normBand(x, ...)
}
\arguments{
\item{x}{A vector of values whose density is to be calculated}

\item{...}{Other arguments to be passed down to \code{sm.density}.}
}
\value{
A named vector of scalar measures of fit
}
\description{
Calculates a kernel density estimate of the data along with confidence bounds.
It also computes a normal density and confidence bounds for the normal density
with the same mean and variance as the observed data.
}
\details{
The function is largely cribbed from the \pkg{sm} package by
Bowman and Azzalini
}
\references{
A.W> Bowman and A. Azzalini, R package sm: nonparametric smoothing methods
(verstion 5.6).
}
\author{
Dave Armstrong, A.W. Bowman and A. Azzalini
}
