% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{print.srr}
\alias{print.srr}
\title{Print Method for Silber, Rosenbaum and Ross Importance Measure}
\usage{
\method{print}{srr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{srr}.}

\item{...}{Other arguments passed down to \code{print}}
}
\value{
Printed output
}
\description{
Prints the results of the \code{srr_imp} function
}
