% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{relig_imp}
\alias{relig_imp}
\title{World Values Survey Religious Importance}
\format{
A data frame with 224 rows and 4 variables
\describe{
  \item{country}{Country of respondent residence.}
  \item{relig_imp}{Response Category for the religious importance variable: Very Important, 
  Rather Important, Not Very Important and Not At All Important.}
  \item{n}{Proportion of observation in each country-response category.}
  \item{l}{The average value of religious importance on the 1-4 scale.}
}
}
\description{
A subset of data from the second thorugh fifth waves of the World Values Survey measuring
religious importance.
}
\details{
These data come from the same source as the \code{wvs} data.  These are aggregated
responses to the question about religious importance by country and religious importance response.  

The  dataset has 224 rows and 4 variables.  The variables are as follows:
}
\references{
Inglehart, R., C. Haerpfer, A. Moreno, C. Welzel, K. Kizilova, J. Diez-Medrano, M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2014a. World Values Survey: Round Two - Country-Pooled Datafile Version: \url{https://www.worldvaluessurvey.org/WVSDocumentationWV2.jsp}. Madrid: JD Systems Institute.

Inglehart, R., C. Haerpfer, A. Moreno, C. Welzel, K. Kizilova, J. Diez-Medrano, M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2014b. World Values Survey: Round Three - Country-Pooled Datafile Version: \url{https://www.worldvaluessurvey.org/WVSDocumentationWV3.jsp}. Madrid: JD Systems Institute.

Inglehart, R., C. Haerpfer, A. Moreno, C. Welzel, K. Kizilova, J. Diez-Medrano, M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2014c. World Values Survey: Round Four - Country-Pooled Datafile Version: \url{https://www.worldvaluessurvey.org/WVSDocumentationWV4.jsp}. Madrid: JD Systems Institute.

Inglehart, R., C. Haerpfer, A. Moreno, C. Welzel, K. Kizilova, J. Diez-Medrano, M. Lagos, P. Norris, E. Ponarin & B. Puranen et al. (eds.). 2014d. World Values Survey: Round Five - Country-Pooled Datafile Version: \url{https://www.worldvaluessurvey.org/WVSDocumentationWV5.jsp}. Madrid: JD Systems Institute.
}
\keyword{datasets}
