% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{rrPlot}
\alias{rrPlot}
\title{Residual-Residual Plot}
\usage{
rrPlot(formula, xlabels = NULL, ylab = NULL, data, return = c("grid", "grobs"))
}
\arguments{
\item{formula}{Formula giving the variables to be plotted.}

\item{xlabels}{Vector of character strings giving the labs of
variables to be used in place of the variable names.}

\item{ylab}{Character string giving y-variable label to be
used instead of variable name.}

\item{data}{A data frame that holds the variables to be plotted.}

\item{return}{A string identify what to return.  If \sQuote{grid}, 
then a \code{cowplot} object is returned with all plots printed.  
If \sQuote{grobs} then a list with all of the individual ggplots/grobs
is returned.}
}
\value{
A \code{cowplot} object.
}
\description{
Produces a linear scatterplot array with marginal histograms.
The plots have OLS regression lines and a 45-degree line.
}
