% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{simple_slopes}
\alias{simple_slopes}
\title{Calculate Simple Slopes}
\usage{
simple_slopes(mod, quant_var, cat_var)
}
\arguments{
\item{mod}{A model object that contains an interaction between a
quantitative variable and a factor.}

\item{quant_var}{A character string giving the name of the quantitative
variable ine the interaction.}

\item{cat_var}{A character string giving the name of the factor
variable ine the interaction.}
}
\value{
A data frame giving the conditional partial effect
along with standard errors, t-statistics and p-values.
}
\description{
Calculates Simple Slopes from an interaction between a categorical
and quantitative variable.
}
