% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{gg_hmf}
\alias{gg_hmf}
\title{Heatmap Fit Plot using GGplot}
\usage{
gg_hmf(observed, prob, span = NULL, nbin = 20, ...)
}
\arguments{
\item{observed}{Vector of observe (0/1) values used in a 
binary regression model.}

\item{prob}{Vector of predicted probabilities from the model 
with \code{observed} as the dependent variable.}

\item{span}{Optional span parameter to be passed in.  If 
\code{NULL}, AICc will be used to find the appropriate 
span for the loess smooth.}

\item{nbin}{Number of bins for the histogram.}

\item{...}{Currently unimplemented.}
}
\value{
Two ggplots - the main heatmap Fit plot and a 
histogram that can be included as a marginal density.
}
\description{
Makes a Heatmap Fit plot (Esary and Pierce, 2012) using
GGPlot rather than lattice that the \code{heatmapFit} package
uses.
}
