% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{print.ss}
\alias{print.ss}
\title{Print Method for Simple Slopes}
\usage{
\method{print}{ss}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ss}.}

\item{...}{Other arguments passed down to \code{print}}
}
\value{
Printed output
}
\description{
Prints the results of the Simple Slopes function
}
