% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_functions.r
\name{srr_imp}
\alias{srr_imp}
\title{Absolute Importance Measure}
\usage{
srr_imp(
  obj,
  data,
  boot = TRUE,
  R = 250,
  level = 0.95,
  ci_method = c("perc", "norm", "bca"),
  ...
)
}
\arguments{
\item{obj}{Model object, must be able to use \code{predict(obj, type="terms")}.}

\item{data}{A data frame used to estiamte the model.}

\item{boot}{Logical indicating whether bootstrap confidence intervals should
be produced and included.}

\item{R}{If \code{boot=TRUE}, the number of bootstrap samples to be used.}

\item{level}{Cofidence level used for the confidence interval.}

\item{ci_method}{Character string giving the method for calculating the
bootstrapped confidence interval.}

\item{...}{Other arguments being passed down to \code{boot}.}
}
\value{
A data frame of importance measures with optimal bootstrapped confidence intervals.
}
\description{
Calculates absolute importance along the lines consistent with
relative importance as defined by Silber, Rosenbaum and Ross (1995)
}
\references{
Silber, J. H., Rosenbaum, P. R. and Ross, R N (1995) Comparing the Contributions of Groups of Predictors: Which Outcomes Vary with Hospital Rather than Patient Characteristics? JASA 90, 7–18.
}
